refactor: proximity system upgrade, refactor audio muffle to use proximity system API
